Isolate napari-workflows tests
